// apps/trip-shorts/types/video.types.ts
export interface Video {
  thumbnail: string;
  title: string;
  duration: string;
  price: string;
  creator: string;
  rating: number;
  views: string;
  likes: string;
  location: string;
  oneLineComment: string;
}

export interface ShortForm {
  id: number;
  thumbnail: string;
  title: string;
  duration: string;
  price: string;
  creator: string;
  rating: number;
  views: string;
  likes: string;
  location: string;
  oneLineComment: string;
}